Add files via upload
This analysis is based on the attrition rate of a company. Attrition is the departure of employees from an organization for any reason (voluntary or involuntary), including resignation, termination, death, or retirement. Attrition rate is the rate at which employees leave an organization divided by the average number of employees at the organization over a given period.

Dataset was provided by @MeriSKILL.

Power Query was used for data cleaning. No duplicate values in the dataset.

For easy analysis, the ages had to be in age groups. I created a new column with the conditional column for the age grouping.

Insights:
There are 1470 employees in the company at the given period

3 departments in the organization.

Data Analysis was done in power BI

To get the attrition rate, Dax measure in Power BI was used on the dataset. The attrition rate is calculated to be 16.12%

Average monthly income is $6.5k. A plot of Male against Female attrition shows more male attrition rate 60% compared to the female at 40%.

The company has three departments, Research and Development tops the list with most employees leaving the company with 133, followed by Sales department 92 people and 12 from Human Resources department.

Research and Development department has the highest number of employees. Employees were group into 7 age groups, those between the age of 30–35years top the list with highest attrition rate, followed by age group 36–41 years. Age groups 54–60 and 18–23 have the lowest attrition rate of all.

Recommendations:
1. Departments with high attrition rates should be investigated and employee retention strategies should be introduced.

2. Provide adequate support and mentoring for job roles with high attrition rate.

3. It was observed that male attrition rate is high compared to female so equality in the workplace should be considered.